Update activitty prices properties

+++ FClub.Models/Models/Activitty.cs
@@ -25,16 +25,16 @@ namespace FClub.Models.Models
 		public ICollection<ActivittyDays> ActivittyDays { get; set; }
 		[Display(Name = "Price per class")]
 	
-		public double PricePerClass { get; set; }
+		public int PricePerClass { get; set; }
 		[Display(Name = "Price per month")]
 
-		public double PricePerMonth { get; set; }
+		public int PricePerMonth { get; set; }
 		[Display(Name = "Price per 3 months")]
 
-		public double PricePer3Months { get; set; }
+		public int PricePer3Months { get; set; }
 		[Display(Name = "Free pass")]
 
-		public double FreePass { get; set; }
+		public int FreePass { get; set; }
 		[Required]
 		
 		public int FromToPeriodId { get; set; }
